move tests to more sensible class, added test for RSA4c3
